+++ src/RcsbFvWeb/RcsbFvModule/RcsbFvEntity.ts
@@ -1,12 +1,12 @@
 import {SequenceReference, Source} from "../../RcsbGraphQL/Types/Borrego/GqlTypes";
 import {RcsbFvAbstractModule} from "./RcsbFvAbstractModule";
-import {RcsbFvModuleBuildInterface, RcsbFvModuleInterface} from "./RcsbFvModuleInterface";
-import {RcsbFv, RcsbFvRowConfigInterface} from "@rcsb/rcsb-saguaro";
+import {RcsbFvModuleBuildInterface} from "./RcsbFvModuleInterface";
+import {RcsbFvRowConfigInterface} from "@rcsb/rcsb-saguaro";
 import {SequenceCollectorDataInterface} from "../CollectTools/SequenceCollector/SequenceCollector";
 
 export class RcsbFvEntity extends RcsbFvAbstractModule {
 
-    public async build(buildConfig: RcsbFvModuleBuildInterface): Promise<RcsbFv> {
+    public async build(buildConfig: RcsbFvModuleBuildInterface): Promise<void> {
         const entityId: string = buildConfig.entityId;
         const source: Array<Source> = buildConfig.additionalConfig?.filters? buildConfig.additionalConfig.sources : [Source.PdbEntity, Source.Uniprot];
         const seqResult:SequenceCollectorDataInterface = await this.sequenceCollector.collect({
@@ -24,10 +24,8 @@ export class RcsbFvEntity extends RcsbFvAbstractModule {
         this.boardConfigData.length = this.sequenceCollector.getSequenceLength();
         this.boardConfigData.includeAxis = true;
         this.rowConfigData = !buildConfig.additionalConfig?.hideAlignments ? seqResult.sequence.concat(seqResult.alignment).concat(annResult) : seqResult.sequence.concat(seqResult.alignment).concat(annResult);
-        this.display();
-        if(buildConfig.resolve!=null)
-            buildConfig.resolve(this);
-        return this.rcsbFv;
+        await this.display();
+        return void 0;
     }
 
 }